+++ resources/migrations/20220120121117-user-mappings.up.sql
@@ -1,0 +1,10 @@
+CREATE TABLE user_mappings (
+  userId varchar(255) NOT NULL,
+  extIdAttribute varchar(255) NOT NULL,
+  extIdValue varchar(255) NOT NULL,
+  PRIMARY KEY (userId, extIdAttribute, extIdValue),
+  FOREIGN KEY (userId) REFERENCES users
+);
+--;;
+CREATE INDEX index_user_mappings ON user_mappings(extIdAttribute, extIdValue);
+