+++ test/clj/rems/api/test_user_settings.clj
@@ -9,7 +9,7 @@
             [rems.db.user-settings :as user-settings]
             [rems.handler :refer [handler]]
             [rems.json :as json]
-            [rems.testing-util :refer [with-fixed-time]]
+            [rems.testing-util :refer [with-fixed-time with-fake-login-users]]
             [ring.mock.request :refer :all]
             [stub-http.core :as stub])
   (:import [java.util UUID]))
@@ -106,28 +106,29 @@
                  read-body-and-status))))
 
     (testing "success"
-      (with-fixed-time (time-core/date-time 2021)
-        (fn []
-          (run-with-ega-server
-           {"/p/api_key/generate" {:status 200 :content-type "application/json" :body (json/generate-string {:token "access-token-xyz"})}}
-           (fn [_server]
-             (let [cookie (login-with-cookies user-id)
-                   csrf (get-csrf-token cookie)]
-               (testing "generate-ega-api-key with session"
-                 (is (= {:status 200
-                         :body {:success true
-                                :api-key-expiration-date "2022-01-01T00:00:00.000Z"}} ; one year expiry
-                        (-> (request :post "/api/user-settings/generate-ega-api-key")
-                            (header "Cookie" cookie)
-                            (header "x-csrf-token" csrf)
-                            handler
-                            assert-response-is-ok
-                            read-body-and-status)))
-                 (is (= (time-core/date-time 2022)
-                        (get-in (user-settings/get-user-settings user-id) [:ega :api-key-expiration-date]))
-                     "one year expiry")
-                 (is (= {:ega {:api-key "access-token-xyz"}}
-                        (user-secrets/get-user-secrets user-id))))))))))))
+      (with-fake-login-users {user-id {:sub user-id}}
+        (with-fixed-time (time-core/date-time 2021)
+          (fn []
+            (run-with-ega-server
+             {"/p/api_key/generate" {:status 200 :content-type "application/json" :body (json/generate-string {:token "access-token-xyz"})}}
+             (fn [_server]
+               (let [cookie (login-with-cookies user-id)
+                     csrf (get-csrf-token cookie)]
+                 (testing "generate-ega-api-key with session"
+                   (is (= {:status 200
+                           :body {:success true
+                                  :api-key-expiration-date "2022-01-01T00:00:00.000Z"}} ; one year expiry
+                          (-> (request :post "/api/user-settings/generate-ega-api-key")
+                              (header "Cookie" cookie)
+                              (header "x-csrf-token" csrf)
+                              handler
+                              assert-response-is-ok
+                              read-body-and-status)))
+                   (is (= (time-core/date-time 2022)
+                          (get-in (user-settings/get-user-settings user-id) [:ega :api-key-expiration-date]))
+                       "one year expiry")
+                   (is (= {:ega {:api-key "access-token-xyz"}}
+                          (user-secrets/get-user-secrets user-id)))))))))))))
 
 (deftest test-delete-api-key
   (test-data/create-test-api-key!)
@@ -138,23 +139,24 @@
     (test-helpers/create-workflow! {:handlers [handler-id]})
 
     (testing "setup api-key to delete"
-      (with-fixed-time (time-core/date-time 2021)
-        (fn []
-          (run-with-ega-server
-           {"/p/api_key/generate" {:status 200 :content-type "application/json" :body (json/generate-string {:token "access-token-xyz"})}}
-           (fn [_server]
-             (let [cookie (login-with-cookies handler-id)
-                   csrf (get-csrf-token cookie)]
-               (testing "generate-ega-api-key with session"
-                 (is (= {:status 200
-                         :body {:success true
-                                :api-key-expiration-date "2022-01-01T00:00:00.000Z"}}
-                        (-> (request :post "/api/user-settings/generate-ega-api-key")
-                            (header "Cookie" cookie)
-                            (header "x-csrf-token" csrf)
-                            handler
-                            assert-response-is-ok
-                            read-body-and-status))))))))))
+      (with-fake-login-users {handler-id {:sub handler-id}}
+        (with-fixed-time (time-core/date-time 2021)
+          (fn []
+            (run-with-ega-server
+             {"/p/api_key/generate" {:status 200 :content-type "application/json" :body (json/generate-string {:token "access-token-xyz"})}}
+             (fn [_server]
+               (let [cookie (login-with-cookies handler-id)
+                     csrf (get-csrf-token cookie)]
+                 (testing "generate-ega-api-key with session"
+                   (is (= {:status 200
+                           :body {:success true
+                                  :api-key-expiration-date "2022-01-01T00:00:00.000Z"}}
+                          (-> (request :post "/api/user-settings/generate-ega-api-key")
+                              (header "Cookie" cookie)
+                              (header "x-csrf-token" csrf)
+                              handler
+                              assert-response-is-ok
+                              read-body-and-status)))))))))))
 
     (testing "without authentication"
       (let [{:keys [body] :as response} (-> (request :post "/api/user-settings/delete-ega-api-key")
@@ -182,19 +184,21 @@
                  read-body-and-status))))
 
     (testing "success"
-      (with-fixed-time (time-core/date-time 2021)
-        (fn []
-          (run-with-ega-server
-           {{:path (str "/p/api_key/" handler-id) :method "DELETE"} {:status 200 :content-type "application/json" :body ""}}
-           (fn [server]
-             (let [cookie (login-with-cookies handler-id)
-                   csrf (get-csrf-token cookie)]
-               (testing "delete-ega-api-key with session"
-                 (is (= {:status 200
-                         :body {:success true}}
-                        (-> (request :post "/api/user-settings/delete-ega-api-key")
-                            (header "Cookie" cookie)
-                            (header "x-csrf-token" csrf)
-                            handler
-                            assert-response-is-ok
-                            read-body-and-status))))))))))))
+      (with-fake-login-users {handler-id {:sub handler-id}}
+        (with-fixed-time (time-core/date-time 2021)
+          (fn []
+            (run-with-ega-server
+             {{:path (str "/p/api_key/" handler-id) :method "DELETE"} {:status 200 :content-type "application/json" :body ""}}
+             (fn [server]
+               (let [cookie (login-with-cookies handler-id)
+                     csrf (get-csrf-token cookie)]
+                 (testing "delete-ega-api-key with session"
+                   (is (= {:status 200
+                           :body {:success true}}
+                          (-> (request :post "/api/user-settings/delete-ega-api-key")
+                              (header "Cookie" cookie)
+                              (header "x-csrf-token" csrf)
+                              handler
+                              assert-response-is-ok
+                              read-body-and-status)))))))))))))
+